# Commander/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TeamServer",
    "version": "v1"
  },
  "paths": {
...
    "/Implants/{implantId}": {
      "get": {
        "tags": [
          "Implants"
        ],
        "parameters": [
          {
            "name": "implantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "withData",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Implants"
        ],
        "parameters": [
          {
            "name": "implantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}